refactor: add reveal module and clean utils module up
